Reorganized include directories
