+++ Ctrl 2K Engine 0.0.1/Ctrl 2K Engine 0.0.1.vcxproj
@@ -70,16 +70,16 @@
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <TargetName>$(ProjectName)d</TargetName>
-    <IntDir>$(SolutionDir)\$(Platform)\Bin\</IntDir>
+    <IntDir>$(ProjectDir)\Bin\$(Platform)\</IntDir>
     <IncludePath>$(ProjectDir)\Dependencies\SDL2_mixer-2.0.2\include;$(ProjectDir)\Include;$(ProjectDir)\Dependencies\SDL2-2.0.7\include;$(ProjectDir)\Dependencies\SOIL for opengl 3.0 context and above\src;$(ProjectDir)\Dependencies\freeglut-3.0.0\include;$(ProjectDir)\Dependencies\glew-2.0.0\include;$(IncludePath)</IncludePath>
     <LibraryPath>$(ProjectDir)\Dependencies\SDL2_mixer-2.0.2\lib\x86;$(ProjectDir)\Dependencies\SOIL for opengl 3.0 context and above\lib;$(ProjectDir)\Dependencies\SDL2-2.0.7\lib\x86;$(ProjectDir)\Dependencies\freeglut-3.0.0\lib;$(ProjectDir)\Dependencies\glew-2.0.0\lib\Debug\Win32;$(LibraryPath)</LibraryPath>
-    <OutDir>$(SolutionDir)\$(Platform)\$(Configuration)\</OutDir>
+    <OutDir>$(SolutionDir)\$(Platform)\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
-    <IntDir>$(SolutionDir)\$(Platform)\Bin\</IntDir>
+    <IntDir>$(ProjectDir)\Bin\$(Platform)\</IntDir>
     <IncludePath>$(ProjectDir)\Dependencies\SDL2_mixer-2.0.2\include;$(ProjectDir)\Include;$(ProjectDir)\Dependencies\SDL2-2.0.7\include;$(ProjectDir)\Dependencies\SOIL for opengl 3.0 context and above\src;$(ProjectDir)\Dependencies\freeglut-3.0.0\include;$(ProjectDir)\Dependencies\glew-2.0.0\include;$(IncludePath)</IncludePath>
     <LibraryPath>$(ProjectDir)\Dependencies\SDL2_mixer-2.0.2\lib\x86;$(ProjectDir)\Dependencies\SOIL for opengl 3.0 context and above\lib;$(ProjectDir)\Dependencies\SDL2-2.0.7\lib\x86;$(ProjectDir)\Dependencies\freeglut-3.0.0\lib;$(ProjectDir)\Dependencies\glew-2.0.0\lib\Release\Win32;$(LibraryPath)</LibraryPath>
-    <OutDir>$(SolutionDir)\$(Platform)\$(Configuration)\</OutDir>
+    <OutDir>$(SolutionDir)\$(Platform)\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <IncludePath>$(ProjectDir)\Dependencies\SDL2_mixer-2.0.2\include;$(ProjectDir)\Include;$(ProjectDir)\Dependencies\SDL2-2.0.7\include;$(ProjectDir)\Dependencies\SOIL for opengl 3.0 context and above\src;$(ProjectDir)\Dependencies\freeglut-3.0.0\include;$(ProjectDir)\Dependencies\glew-2.0.0\include;$(IncludePath)</IncludePath>
