+++ Ctrl Test.sln
@@ -1,0 +1,81 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29306.81
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{9195EE2C-80D2-44E4-8B7F-BDA29544780B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tools", "Tools", "{394EB5CA-5D00-4573-A6FF-97C979310C58}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GreatBath", "Core\GreatBath\GreatBath.vcxproj", "{55BB3977-E4AA-4C08-B457-B1214977396C}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Petra", "Core\Petra\Petra.vcxproj", "{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SunDial", "Core\SunDial\SunDial.vcxproj", "{EB11A967-DAB8-4DE2-894C-F5014E62137E}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SunDial Profiler", "Tools\SunDial Profiler\SunDial Profiler.vcxproj", "{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CtrlTest", "Ctrl Test\CtrlTest.vcxproj", "{0ED39A78-FF6C-411B-895F-BF8953C6C533}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x64.ActiveCfg = Debug|x64
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x64.Build.0 = Debug|x64
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x86.ActiveCfg = Debug|Win32
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x86.Build.0 = Debug|Win32
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x64.ActiveCfg = Release|x64
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x64.Build.0 = Release|x64
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x86.ActiveCfg = Release|Win32
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x86.Build.0 = Release|Win32
+		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Debug|x64.ActiveCfg = Debug|x64
+		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Debug|x64.Build.0 = Debug|x64
+		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Debug|x86.ActiveCfg = Debug|Win32
+		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Debug|x86.Build.0 = Debug|Win32
+		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Release|x64.ActiveCfg = Release|x64
+		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Release|x64.Build.0 = Release|x64
+		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Release|x86.ActiveCfg = Release|Win32
+		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Release|x86.Build.0 = Release|Win32
+		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Debug|x64.ActiveCfg = Debug|x64
+		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Debug|x64.Build.0 = Debug|x64
+		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Debug|x86.ActiveCfg = Debug|Win32
+		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Debug|x86.Build.0 = Debug|Win32
+		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Release|x64.ActiveCfg = Release|x64
+		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Release|x64.Build.0 = Release|x64
+		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Release|x86.ActiveCfg = Release|Win32
+		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Release|x86.Build.0 = Release|Win32
+		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Debug|x64.ActiveCfg = Debug|x64
+		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Debug|x64.Build.0 = Debug|x64
+		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Debug|x86.ActiveCfg = Debug|Win32
+		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Debug|x86.Build.0 = Debug|Win32
+		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Release|x64.ActiveCfg = Release|x64
+		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Release|x64.Build.0 = Release|x64
+		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Release|x86.ActiveCfg = Release|Win32
+		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Release|x86.Build.0 = Release|Win32
+		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Debug|x64.ActiveCfg = Debug|x64
+		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Debug|x64.Build.0 = Debug|x64
+		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Debug|x86.ActiveCfg = Debug|Win32
+		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Debug|x86.Build.0 = Debug|Win32
+		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Release|x64.ActiveCfg = Release|x64
+		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Release|x64.Build.0 = Release|x64
+		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Release|x86.ActiveCfg = Release|Win32
+		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{55BB3977-E4AA-4C08-B457-B1214977396C} = {9195EE2C-80D2-44E4-8B7F-BDA29544780B}
+		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF} = {9195EE2C-80D2-44E4-8B7F-BDA29544780B}
+		{EB11A967-DAB8-4DE2-894C-F5014E62137E} = {9195EE2C-80D2-44E4-8B7F-BDA29544780B}
+		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1} = {394EB5CA-5D00-4573-A6FF-97C979310C58}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {6DDFDF4A-2092-4AAE-9196-38FF189B5401}
+	EndGlobalSection
+EndGlobal
