--- Ctrl test/Ctrl test.sln
@@ -1,61 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29306.81
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CtrlTest", "CtrlTest.vcxproj", "{0ED39A78-FF6C-411B-895F-BF8953C6C533}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GreatBath", "..\Ctrl 2K Engine 0.0.1\GreatBath.vcxproj", "{55BB3977-E4AA-4C08-B457-B1214977396C}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SunDial", "..\Ctrl 2K Engine 0.0.1\SunDial.vcxproj", "{EB11A967-DAB8-4DE2-894C-F5014E62137E}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SunDial Profiler", "..\Ctrl 2K Engine 0.0.1\SunDial Profiler.vcxproj", "{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Debug|x64.ActiveCfg = Debug|x64
-		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Debug|x64.Build.0 = Debug|x64
-		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Debug|x86.ActiveCfg = Debug|Win32
-		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Debug|x86.Build.0 = Debug|Win32
-		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Release|x64.ActiveCfg = Release|x64
-		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Release|x64.Build.0 = Release|x64
-		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Release|x86.ActiveCfg = Release|Win32
-		{0ED39A78-FF6C-411B-895F-BF8953C6C533}.Release|x86.Build.0 = Release|Win32
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x64.ActiveCfg = Debug|x64
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x64.Build.0 = Debug|x64
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x86.ActiveCfg = Debug|Win32
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x86.Build.0 = Debug|Win32
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x64.ActiveCfg = Release|x64
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x64.Build.0 = Release|x64
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x86.ActiveCfg = Release|Win32
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x86.Build.0 = Release|Win32
-		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Debug|x64.ActiveCfg = Debug|x64
-		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Debug|x64.Build.0 = Debug|x64
-		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Debug|x86.ActiveCfg = Debug|Win32
-		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Debug|x86.Build.0 = Debug|Win32
-		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Release|x64.ActiveCfg = Release|x64
-		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Release|x64.Build.0 = Release|x64
-		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Release|x86.ActiveCfg = Release|Win32
-		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Release|x86.Build.0 = Release|Win32
-		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Debug|x64.ActiveCfg = Debug|x64
-		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Debug|x64.Build.0 = Debug|x64
-		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Debug|x86.ActiveCfg = Debug|Win32
-		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Debug|x86.Build.0 = Debug|Win32
-		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Release|x64.ActiveCfg = Release|x64
-		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Release|x64.Build.0 = Release|x64
-		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Release|x86.ActiveCfg = Release|Win32
-		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6DDFDF4A-2092-4AAE-9196-38FF189B5401}
-	EndGlobalSection
-EndGlobal
