+++ Ctrl test/Source/CtrlTest.cpp
@@ -6,9 +6,9 @@
 #include <vector>
 #include <assert.h>
 
-#include "Logger.h"
-#include "EngineClock.h"
-#include "Profiler.h"
+#include "Core/GreatBath/Logger.h"
+#include "Core/SunDial/EngineClock.h"
+#include "Tools/SunDial Profiler/Profiler.h"
 
 #define CTRLTEST_ERROR std::this_thread::sleep_for(std::chrono::seconds(1)); assert(false);
