+++ Tools/SunDial Profiler/Profiler.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <chrono>
-#include "../../../Core/GreatBath/FilePathProcessing.h"
+#include "Core/GreatBath/FilePathProcessing.h"
 
 class SunDialProfiler
 {
